pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "arabic-witness-pipeline"
version = "0.1.0"
description = "Arabic manuscript witness collation pipeline"
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10.0",
  "numpy>=1.26",
  "streamlit>=1.35",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
